@@ -27,28 +27,31 @@
     "format": "prettier --write ."
   },
   "devDependencies": {
-    "eslint": "^8.52.0",
+    "@eslint/eslintrc": "^3.3.0",
+    "@eslint/js": "^9.21.0",
+    "@sveltejs/vite-plugin-svelte": "^5.0.3",
+    "@tsconfig/svelte": "^5.0.4",
+    "@types/ua-parser-js": "^0.7.36",
+    "@typescript-eslint/eslint-plugin": "^8.25.0",
+    "eslint": "^9.21.0",
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-prettier": "^5.0.1",
-    "eslint-plugin-svelte": "^2.34.1",
+    "eslint-plugin-svelte": "^3.0.2",
+    "globals": "^16.0.0",
     "prettier": "^3.1.0",
     "prettier-plugin-svelte": "^3.1.0",
-    "svelte": "^3.54.0",
-    "svelte-check": "^2.10.0",
-    "svelte-eslint-parser": "^0.33.1",
-    "@sveltejs/vite-plugin-svelte": "^2.0.0",
-    "@tsconfig/svelte": "^3.0.0",
+    "svelte": "~5.20.2",
+    "svelte-check": "^4.1.4",
     "tslib": "^2.4.1",
-    "typescript": "^4.9.3",
-    "@typescript-eslint/eslint-plugin": "^6.9.1",
-    "@types/ua-parser-js": "^0.7.36",
-    "vite": "^4.0.0",
-    "vite-plugin-dts": "^1.7.1",
+    "typescript": "~5.7.2",
+    "vite": "^6.2.0",
+    "vite-plugin-dts": "^4.5.0",
     "vite-plugin-top-level-await": "^1.2.2",
     "vite-plugin-wasm": "^3.1.0"
   },
   "dependencies": {
     "rxjs": "^6.6.7",
+    "svelte-eslint-parser": "^1.0.0",
     "ua-parser-js": "^1.0.33"
   }
 }
